refactor(@dpc-sdp/ripple-tide-publication): :recycle: combine publication and publication-page
