+++ examples/nuxt-app/package.json
@@ -23,7 +23,6 @@
     "@dpc-sdp/ripple-tide-media": "2.1.0-alpha.37",
     "@dpc-sdp/ripple-tide-news": "2.1.0-alpha.37",
     "@dpc-sdp/ripple-tide-publication": "2.1.0-alpha.38",
-    "@dpc-sdp/ripple-tide-publication-page": "2.1.0-alpha.38",
     "@dpc-sdp/ripple-ui-core": "2.1.0-alpha.36",
     "@dpc-sdp/ripple-ui-forms": "2.1.0-alpha.39",
     "dotenv": "^16.0.3"
