+++ packages/ripple-tide-api/src/ripple-modules.ts
@@ -5,7 +5,6 @@ export default [
   '@dpc-sdp/ripple-tide-media/nuxt',
   '@dpc-sdp/ripple-tide-news/nuxt',
   '@dpc-sdp/ripple-tide-publication/nuxt',
-  '@dpc-sdp/ripple-tide-publication-page/nuxt',
   '@dpc-sdp/ripple-ui-core/nuxt',
   '@dpc-sdp/ripple-ui-forms/nuxt'
 ]
