--- packages/ripple-tide-publication-page/CHANGELOG.md
@@ -1,132 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [2.1.0-alpha.38](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.37...v2.1.0-alpha.38) (2022-12-14)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.37](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.36...v2.1.0-alpha.37) (2022-12-14)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.36](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.35...v2.1.0-alpha.36) (2022-12-14)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.35](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.34...v2.1.0-alpha.35) (2022-12-14)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.33](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.32...v2.1.0-alpha.33) (2022-12-13)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.31](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.30...v2.1.0-alpha.31) (2022-12-09)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.28](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.27...v2.1.0-alpha.28) (2022-12-07)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.27](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.26...v2.1.0-alpha.27) (2022-12-06)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.26](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.25...v2.1.0-alpha.26) (2022-12-06)
-
-
-### Features
-
-* **@dpc-sdp/ripple-tide-event:** :sparkles: add event content type ([e079cd5](https://github.com/dpc-sdp/ripple-framework/commit/e079cd5580274f57ec9d082c81a312e349048531))
-
-
-
-
-
-# [2.1.0-alpha.25](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.24...v2.1.0-alpha.25) (2022-12-05)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.24](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.23...v2.1.0-alpha.24) (2022-12-02)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.22](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.21...v2.1.0-alpha.22) (2022-12-01)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.21](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.20...v2.1.0-alpha.21) (2022-11-30)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.20](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.19...v2.1.0-alpha.20) (2022-11-29)
-
-**Note:** Version bump only for package @dpc-sdp/ripple-tide-publication-page
-
-
-
-
-
-# [2.1.0-alpha.19](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.18...v2.1.0-alpha.19) (2022-11-28)
-
-
-### Bug Fixes
-
-* **@dpc-sdp/ripple-tide-publication:** :bug: post-merge fixes ([d8f5207](https://github.com/dpc-sdp/ripple-framework/commit/d8f52074f42dcbf325f2c62a2ed99f8829182801))
-
-
-### Features
-
-* **@dpc-sdp/ripple-tide-api:** :sparkles: add publication index api, mapping ([8d393e7](https://github.com/dpc-sdp/ripple-framework/commit/8d393e78dad09bb617e0979f8efd9620ee79bcbe))
-* **@dpc-sdp/ripple-tide-publication:** :sparkles: add child content type ([8f0c6ce](https://github.com/dpc-sdp/ripple-framework/commit/8f0c6ceaba2abae282261bbf39c01d7f417de42c))
