--- packages/ripple-tide-publication-page/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@dpc-sdp/ripple-tide-publication-page",
-  "version": "2.1.0-alpha.38",
-  "description": "Ripple mappings and components for Tide PublicationPage Content type",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "type": "module",
-  "exports": {
-    ".": "./dist/index.js",
-    "./component": "./src/index.vue",
-    "./types": "./src/types.ts",
-    "./nuxt": "./src/nuxt.ts"
-  },
-  "license": "MIT",
-  "keywords": [],
-  "scripts": {
-    "clean": "rimraf dist* && rimraf tsconfig.tsbuildinfo",
-    "build": "yarn clean && tsc -p tsconfig.json",
-    "watch": "yarn clean && tsc -p tsconfig.json -w"
-  },
-  "dependencies": {
-    "@dpc-sdp/ripple-tide-api": "2.1.0-alpha.37",
-    "@dpc-sdp/ripple-tide-publication": "2.1.0-alpha.38"
-  },
-  "files": [
-    "dist",
-    "src",
-    "!**/*.test.*",
-    "!**/*.json",
-    "CHANGELOG.md",
-    "LICENSE",
-    "README.md"
-  ]
-}
