--- packages/ripple-tide-publication-page/src/components/tide-publication-page-header.vue
@@ -1,25 +1,0 @@
-<script lang="ts">
-export default { name: 'TidePublicationPageHeader' }
-</script>
-
-<template>
-  <RplHeaderHeroHeader
-    :corner-top="true"
-    :corner-bottom="true"
-    :behind-nav="true"
-    :breadcrumbs="true"
-    :title="header.title"
-  >
-    <p class="rpl-type-p-large">{{ header.summary }}</p>
-  </RplHeaderHeroHeader>
-</template>
-
-<script setup lang="ts">
-import { TidePublicationPageHeader } from '../types'
-
-interface Props {
-  header: TidePublicationPageHeader
-}
-
-defineProps<Props>()
-</script>
