--- packages/ripple-tide-publication-page/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "noImplicitAny": false,
-    "declaration": false,
-    "jsx": "preserve",
-    "sourceMap": true,
-    "paths": {
-      "@dpc-sdp/ripple-tide-api": ["./../ripple-tide-api/dist/index.js"]
-    }
-  },
-  "include": ["./src"]
-}
