+++ packages/ripple-tide-publication/src/pages/publication-page.vue
@@ -1,5 +1,7 @@
 <script lang="ts">
-export default { name: 'TidePublicationPage' }
+export default {
+  name: 'TidePublicationPagePage'
+}
 </script>
 
 <template>
@@ -30,15 +32,13 @@ export default { name: 'TidePublicationPage' }
         v-if="page.showInPageNav"
         :headingLevel="page.inPageNavHeadingLevel"
         :components="page.dynamicComponents"
-        class="rpl-u-margin-b-9"
       />
       <TidePublicationBody
-        :details="page.details"
         :components="page.dynamicComponents"
       ></TidePublicationBody>
-      <TidePublicationChapters
-        :chapters="page.chapters"
-      ></TidePublicationChapters>
+      <TidePublicationPagination
+        :pagination="page.publication.pagination"
+      ></TidePublicationPagination>
     </template>
     <template #sidebar>
       <TidePublicationSidebar
@@ -53,12 +53,11 @@ export default { name: 'TidePublicationPage' }
 </template>
 
 <script setup lang="ts">
-import { TideSiteData } from '@dpc-sdp/ripple-tide-api/types'
-import type { TidePublicationPage } from './types'
+import { TidePublicationPagePage } from '../types'
 
 interface Props {
-  site: TideSiteData
-  page: TidePublicationPage
+  site: any
+  page: TidePublicationPagePage
 }
 
 defineProps<Props>()
